converts bin type

+++ src/pmmoto/domain_generation/rdf.pxd
@@ -8,10 +8,10 @@ from .particles.atoms cimport AtomList
 
 cdef extern from "rdf.hpp":
 
-    cdef vector[long int] generate_rdf(
+    cdef vector[unsigned long long] generate_rdf(
         shared_ptr[AtomList] probe_atoms,
         shared_ptr[AtomList] atoms,
         double max_distance,
-        vector[long int] bins,
+        vector[unsigned long long] bins,
         double bin_width
         )
